Unroll into loops and deal with the case where we get a hit at the end of one curve and the start of the following one
Occasionally a floating point error can result in points being returned twice which causes us to count the same intersection twice.

It's tricky to work around this with iterators as it requires inspecting values either side of the iteration.

There's an edge case where this happens on the very first and last points in the path which is not handled here yet.