Fix issue: problem is a threshold issue when we have a collision that's almost but not quite on a collinear section
Assuming 'close' points are the same as the start/end point when we've already found a collinear line for now.
I think an improved version of this could be to just assume the last collision along the edges around the collinear
line are the start and end points, would be a bit less likely to remove more than one point by mistake.