First pass at fixing the glancing corners bug
Can re-cast rays when we might have a glancing collision as they are rare.

This doesn't actually work because the case I've got is actually the same rounding issue but on the following edge
instead of the preceding one. So we need to look for that too, but we should be able to test this by reversing the
path.