Trick to deciding whether or not to subdivide is whether or not the arch converges to a center point or not
This subdivides too much though if we set it precise so we need a better algorithm.

I think the problem is we generate single-inflection point curves sometimes: should be possible to determine a better subdivision point using the canonical form?

Also we're calculating the intersection point twice here. Tests are failing more because the error calculator doesn't work very well than the curves are actually bad.