Add a routine to count the collisions at a particular edge
This is mainly useful for testing but may be useful for implementing other algorithms. Looking for an issue where the edge re-ordering algorithm is triggered in incorrect circumstances. (`set_edge_kinds_by_ray_casting` is more efficient than this as it'll set the edge kinds of everything the ray crosses)