Add some extra collision tests with the rectangle path going in opposite directions
I'm kind of suspicious about what ordered_ray_collisions is doing here: essentially we need the ray to enter the earlier shape first and the later shape last. The comparison here reverses the ordering returned by ray_collisions if one of the edges is anti-clockwise.

(Possibly need to reverse ordering if any edge is anti-clockwise? There's still a possible issue for overlapping interior portions of the shape here)