Break things: remove the re-ordering function entirely
Putting this in its own function highlighted how the re-ordering we were doing didn't really make any sense (ray_collisions() already knows how to order collisions so re-ordering later seems weird)

The problem is deciding what to do when a ray crosses multiple overlapping edges. They need to be ordered based on whether or not the ray is currently inside or outside the path: when the ray is initially outside the path, the earlier edges need to be hit first. When the ray is initially inside the path, the later edges need to be hit first.

We can sort of know that by looking at normals and path directions, but given that the edge-crossing algorithm is programmable this will forever be buggy to a greater or lesser extent, so we need to process overlapped collisions as an exception later on (as we were originally) rather than try to order them without knowing what the algorithm considers 'inside' or 'outside'.