SMALL_DISTANCE test in here needs to use .abs() values
This previously very rarely caused issues as you'd need a ray hitting the same edge twice in succession and both the dx and dy values had to be decreasing, and originally we set the edges in groups individually so it'd set to exterior properly in any case. The new code deals with groups as a whole so a suitable ray that entered and left via a single edge would incorrectly mark the edge as exterior if this issue was hit. 

Rarity of this situation was still quite high, given that only one test failed...